Fix timezones, detailed energy summary improvement
The data is always supplied as Europe/Amsterdam. Make sure to interpret the
data like this as well, otherwise data is visualised with an incorrect offset
of 1 or 2 hours depending on what the difference is between UTC and
Europe/Amsterdam.

Next to that, improve visualisations with running totals to show more info in
one glance.